🐛 Make service cleanup more reliable
JIRA: https://issues.jboss.org/browse/RHMAP-15990

There's a bunch of stuff that previousl would happen in the 'cleanup'
function, and in some cases an exception could be thrown. To ensure
that all resources are scaled down and cleaned up, we basically need
to wrap each step there in a `try...finally`.

This solution is made more complex because we don't really know how
many services are needed, so before this change we'd loop over them
and clean them all up. If we want to chain a bunch of
`try...finally`'s, then we can build that chain in a series of nested
closures at runtime (call `buildingClosure` in a loop), and then
execute all together (`builtFunction()`).

That's what I've attempted to achieve in this solution, and I think I
have. Since we only support MongoDB currently, I haven't been able to
test with more than a single service request. Additionally, where we
were getting parallelization in the service creation/scaling, that's
now lost if we go with this approach. I'm sure there's another way
where we get the best of both worlds, but I haven't gotten there yet.